refactor(anyharness): split live session runtime concerns
