+++ anyharness/crates/anyharness-lib/src/live/sessions/actor/interactions/mod.rs
@@ -1,3 +1,5 @@
 pub mod cleanup;
 pub mod handle;
+pub mod outcomes;
+pub mod plan_decisions;
 pub mod plan_links;
