+++ anyharness/crates/anyharness-lib/src/live/sessions/manager/replay.rs
@@ -1,0 +1,61 @@
+use std::sync::Arc;
+
+use anyharness_contract::v1::SessionEventEnvelope;
+use tokio::sync::broadcast;
+
+use super::LiveSessionManager;
+use crate::live::sessions::actor::spawn::ActorReadyResult;
+use crate::live::sessions::handle::LiveSessionHandle;
+use crate::live::sessions::replay::{spawn_replay_actor, ReplayActorConfig};
+use crate::sessions::model::SessionRecord;
+use crate::sessions::store::SessionStore;
+
+impl LiveSessionManager {
+    pub async fn start_replay_session(
+        &self,
+        session: SessionRecord,
+        events: Vec<SessionEventEnvelope>,
+        speed: f32,
+        session_store: SessionStore,
+        last_seq: i64,
+    ) -> anyhow::Result<(Arc<LiveSessionHandle>, ActorReadyResult)> {
+        let session_id = session.id.clone();
+        let mut sessions = self.live_sessions.write().await;
+        if let Some(existing) = sessions.get(&session_id) {
+            return Ok((
+                existing.clone(),
+                ActorReadyResult {
+                    native_session_id: existing
+                        .native_session_id()
+                        .or(session.native_session_id)
+                        .unwrap_or_default(),
+                },
+            ));
+        }
+
+        let (event_tx, _) = broadcast::channel::<SessionEventEnvelope>(4096);
+        let live_sessions = self.live_sessions.clone();
+        let exit_session_id = session_id.clone();
+        let exit_store = session_store.clone();
+        let on_exit: Box<dyn FnOnce(bool) + Send + 'static> = Box::new(move |errored| {
+            live_sessions.blocking_write().remove(&exit_session_id);
+            if errored {
+                let now = chrono::Utc::now().to_rfc3339();
+                let _ = exit_store.update_status(&exit_session_id, "errored", &now);
+            }
+        });
+
+        let config = ReplayActorConfig {
+            session,
+            events,
+            speed,
+            event_tx,
+            session_store,
+            last_seq,
+            on_exit: Some(on_exit),
+        };
+        let (handle, ready) = spawn_replay_actor(config)?;
+        sessions.insert(session_id, handle.clone());
+        Ok((handle, ready))
+    }
+}
