+++ scripts/max_lines_allowlist.txt
@@ -24,10 +24,6 @@ anyharness/crates/anyharness-lib/src/domains/mobility/service.rs 1534 existing A
 anyharness/crates/anyharness-lib/src/domains/plans/service.rs 653 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/reviews/service/service_tests.rs 1519 existing AnyHarness oversized test file
 anyharness/crates/anyharness-lib/src/domains/runtime_config/service.rs 1057 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/live/sessions/actor/interactions/handle.rs 615 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/live/sessions/background_work/claude.rs 663 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/live/sessions/connection/runtime_client.rs 695 existing AnyHarness oversized file
-anyharness/crates/anyharness-lib/src/live/sessions/manager.rs 933 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/sessions/live_config.rs 675 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/sessions/prompt.rs 1423 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/sessions/service.rs 645 existing AnyHarness oversized file
